requirements.txt
aerich==0.6.3
asyncpg==0.26.0
envparse >= 0.2.0
loguru >= 0.4.4
numpy
orjson
pillow
pytrovich
pytz
tortoise-orm==0.19.2
vkbottle == 4.3.4
vkbottle-types==5.131.146.5
